updated man printf

+++ binary.c
@@ -9,7 +9,7 @@
 int use_b(va_list arg)
 {
 	/*declare local variables*/
-	int divisor = 2, len = 0, i, j, k, newval;
+	long int divisor = 2, len = 0, i, j, k, newval;
 	int *ptr;
 	int val = va_arg(arg, int);
